pyproject.toml
[project]
name = "llm-customization-ops"
version = "0.1.0"
description = "Production-grade LLM customization pipelines (LoRA/QLoRA/DPO) with serving and observability."
requires-python = ">=3.11"
readme = "README.md"
license = {file = "LICENSE"}
authors = [{name = "LLM Customization Ops"}]
dependencies = [
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "typer>=0.12",
  "rich>=13.7",
  "structlog>=24.1",
  "python-json-logger>=2.0",
  "datasets>=2.18",
  "transformers>=4.40",
  "peft>=0.10",
  "trl>=0.8",
  "accelerate>=0.29",
  "torch>=2.2",
  "fastapi>=0.110",
  "uvicorn>=0.29",
  "prometheus-client>=0.20",
  "opentelemetry-api>=1.24",
  "opentelemetry-sdk>=1.24",
  "opentelemetry-instrumentation-fastapi>=0.45b0",
  "opentelemetry-exporter-otlp>=1.24",
  "httpx>=0.27",
  "numpy>=1.26",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
train = [
  "bitsandbytes>=0.43; platform_system != 'Darwin'",
  "sentencepiece>=0.2",
]
serve = [
  "uvicorn[standard]>=0.29",
]
dev = [
  "pytest>=8.1",
  "pytest-cov>=5.0",
  "ruff>=0.4.4",
  "black>=24.3",
  "mypy>=1.9",
  "types-PyYAML",
  "pre-commit>=3.7",
]

[project.scripts]
llm-customization-ops = "llm_customization_ops.cli:app"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.11"
warn_unused_ignores = true
warn_redundant_casts = true
no_implicit_optional = true
check_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-q"
markers = ["slow: slow tests requiring optional deps"]
